Try to keep ratio of discarded questions balanced
Questions that have entities as their answer and not literals (dates and
numbers) tend to be filtered out more often, e.g. if a name could not be
mapped to a MID or if the entity mention transfer was not successful.
This commit tries to keep the ratio between entity-answer, date-answer
and number-answer of discarded questions balanced by artificially
discarding quesitons with a lower discard/keep ratio.